refactor: 新旧実装を整理し、最良の実装のみを残す
- StageHunksRefactoredの実装をStageHunksに統合
- 一時的なファイル（stager_new.go、stager_refactored.go）を削除
- 必要な関数（preparePatchData、getCurrentDiff等）をstager.goに移動
- コードの重複を削除し、アーキテクチャを簡潔に整理

これにより、1つの明確な実装のみが残り、コードベースが整理されました。
